Rework
Where to begin..

The player MUST send a packet to the server for a bunch of very
important stuff to happen. This is 70% of the work. I have to work
on the reply back from the server. But the important stuff is happening.

The reply from the server is so the player knows the item will be
removed from the chest etc